Change type of Row::coefficient
- Row::coefficient was changed from size_t to long double.

+++ src/rcspp/graph/row.hpp
@@ -9,6 +9,6 @@ namespace rcspp {
 
 struct Row {
         size_t index;
-        size_t coefficient;
+        long double coefficient;
 };
 }  // namespace rcspp
